+++ runtime/ishtar.vm/runtime/jit/NativeCallData.cs
@@ -1,0 +1,9 @@
+namespace ishtar;
+
+public struct NativeCallData
+{
+    public nint procedure;
+    public long argCount;
+    public nint returnMemoryPointer;
+    public nint argsPointer;
+}
