+++ runtime/ishtar.vm/runtime/jit/NativeCallInfo.cs
@@ -1,0 +1,6 @@
+namespace ishtar;
+
+public class NativeCallInfo
+{
+    public bool retIsPointer { get; set; }
+}
