+++ runtime/ishtar.vm/runtime/jit/RegisterKind.cs
@@ -1,0 +1,9 @@
+namespace ishtar;
+
+public enum RegisterKind
+{
+    GPR64,
+    GPR32,
+    GPR16,
+    GPR8
+}
