+++ runtime/ishtar.vm/runtime/jit/TypeMarshalBox.cs
@@ -1,0 +1,5 @@
+namespace ishtar;
+
+using vein.runtime;
+
+public record TypeMarshalBox(byte size, VeinClass clazz);
