+++ runtime/ishtar.vm/runtime/jit/x64_AssemblerStep.cs
@@ -1,0 +1,16 @@
+namespace ishtar;
+
+using Iced.Intel;
+
+public struct x64_AssemblerStep
+{
+    public InstructionTarget Instruction;
+    public Register Register;
+    public int StackOffset;
+
+    public enum InstructionTarget
+    {
+        push,
+        mov
+    }
+}
